Move SBOM telemetry to its own package
